Wrap-up with camera, stereo_camera, and undistortrectifier
