+++ src/frontend/CMakeLists.txt
@@ -1,7 +1,9 @@
 ### Add source code for stereoVIO
 target_sources(kimera_vio
   PRIVATE
+  "${CMAKE_CURRENT_LIST_DIR}/Camera.cpp"
   "${CMAKE_CURRENT_LIST_DIR}/StereoCamera.cpp"
+  "${CMAKE_CURRENT_LIST_DIR}/UndistorterRectifier.cpp"
   "${CMAKE_CURRENT_LIST_DIR}/CameraParams.cpp"
   "${CMAKE_CURRENT_LIST_DIR}/StereoFrame.cpp"
   "${CMAKE_CURRENT_LIST_DIR}/StereoMatchingParams.cpp"
